Fix the inconsistent event_id/freq_id typing and use correct printf flags
- various parts of the baseband code weren't entirely consistent in types of
  event_id and freq_id values being passed around. These should be "uint64_t"
  and "uint32_t", respectively

- keep time information always as doubles, particularly when calculating in
  nanoseconds

- match the correct printf flags in logging statements to the type of argument
  being printed (for portability, use the macros defined in intttypes.h; e.g.,
  event_id is 64-bit unsigned, and so needs to be printed with `"%" PRIu64`, etc.)

+++ lib/core/basebandRequestManager.cpp
@@ -112,7 +112,7 @@ void basebandRequestManager::status_callback_single_event(const uint64_t event_i
 
 basebandSlice basebandRequestManager::translate_trigger(const int64_t fpga_time0, const int64_t fpga_width,
                                                         const double dm, const double dm_error,
-                                                        const int64_t freq_id,
+                                                        const uint32_t freq_id,
                                                         const double ref_freq_hz) {
     const double freq = FPGA_FREQ0 + FPGA_DELTA_FREQ * freq_id;
     const double freq_inv_sq_diff = (1. / (freq*freq) - 1. / (ref_freq_hz*ref_freq_hz));
